Refactor: 컴포넌트폴더에 dropdown종류별로 컴포넌트 생성
