+++ fe/src/mocks/milestone/data.ts
@@ -1,16 +1,16 @@
 export const mockMilestones = [
   {
-    name: 'string',
+    name: 'be',
     dueDate: '2022-06-20',
-    description: 'string',
+    description: 'be',
     progressRate: 0.33,
     openCount: 10,
     closeCount: 20,
   },
   {
-    name: 'string',
+    name: 'fe',
     dueDate: '2022-06-20',
-    description: 'string',
+    description: 'fe',
     progressRate: 0.55,
     openCount: 5,
     closeCount: 4,
