+++ fe/src/pages/Issues/IssueTableHeader/index.tsx
@@ -1,3 +1,5 @@
+import LabelDropdown from '@/components/LabelDropdown';
+import MilestoneDropdown from '@/components/MilestoneDropdown';
 import { labelQuery } from '@/stores/selector/labelQuery';
 import { mileStoneQuery } from '@/stores/selector/milestoneQuery';
 
@@ -5,18 +7,23 @@ import DropdownContainer from './DropdownContainer';
 import * as S from './style';
 
 const DropdownLabelList = [
-  { id: 'assignee', header: '담당자', selector: labelQuery },
-  { id: 'label', header: '레이블', selector: labelQuery },
-  { id: 'milestone', header: '마일스톤', selector: labelQuery },
-  { id: 'writer', header: '작성자', selector: labelQuery },
+  { id: 'assignee', header: '담당자', DropdownComponent: LabelDropdown },
+  { id: 'label', header: '레이블', DropdownComponent: LabelDropdown },
+  { id: 'milestone', header: '마일스톤', DropdownComponent: MilestoneDropdown },
+  { id: 'writer', header: '작성자', DropdownComponent: MilestoneDropdown },
 ];
 
 export default function IssueTableHeader() {
   return (
     <S.IssueTableHeader>
       <S.DropdownTextButtonContainer>
-        {DropdownLabelList.map(({ id, header, selector }) => (
-          <DropdownContainer key={id} id={id} header={header} selector={selector} />
+        {DropdownLabelList.map(({ id, header, DropdownComponent }) => (
+          <DropdownContainer
+            key={id}
+            id={id}
+            header={header}
+            DropdownComponent={DropdownComponent}
+          />
         ))}
       </S.DropdownTextButtonContainer>
     </S.IssueTableHeader>
